remove camlimages / prepare new io structure
