+++ io/png/dune
@@ -1,0 +1,19 @@
+(library
+ (name Png)
+ (public_name odiff-io.png)
+ (flags
+  (-w -40 -w +26)
+  (:include png_flags.sexp))
+ (foreign_stubs
+  (language c)
+  (names ReadPng)
+  (flags
+   (:include png_c_flags.sexp)))
+ (c_library_flags
+  (:include png_c_library_flags.sexp))
+ (libraries odiff-core WritePng))
+
+(rule
+ (targets png_flags.sexp png_c_flags.sexp png_c_library_flags.sexp)
+ (action
+  (run ../config/discover.exe)))
