+++ io/png_write/dune
@@ -1,0 +1,21 @@
+(library
+ (name WritePng)
+ (public_name odiff-io.png_write)
+ (flags
+  (-w -40 -w +26)
+  (:include png_write_flags.sexp))
+ (foreign_stubs
+  (language c)
+  (names WritePng)
+  (flags
+   (:include png_write_c_flags.sexp)))
+ (c_library_flags
+  (:include png_write_c_library_flags.sexp)))
+
+(rule
+ (targets
+  png_write_flags.sexp
+  png_write_c_flags.sexp
+  png_write_c_library_flags.sexp)
+ (action
+  (run ../config/discover.exe)))
