+++ test/PngTests.re
@@ -1,14 +1,13 @@
 open TestFramework;
 open ODiffIO;
 
-module Diff = Odiff.Diff.MakeDiff(PureC_IO.IO, PureC_IO.IO);
-module AADiff =
-  Odiff.Diff.MakeDiff(PureC_IO_Bigarray.IO, PureC_IO_Bigarray.IO);
+module Diff = Odiff.Diff.MakeDiff(Png.IO, Png.IO);
+module AADiff = Odiff.Diff.MakeDiff(Png.BigarrayIO, Png.BigarrayIO);
 
 describe("Png comparing", ({test, _}) => {
   test("finds difference between 2 images", ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/orange.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/orange_changed.png");
+    let img1 = Png.IO.loadImage("test/test-images/orange.png");
+    let img2 = Png.IO.loadImage("test/test-images/orange_changed.png");
 
     let (_, diffPixels, diffPercentage) = Diff.compare(img1, img2, ());
 
@@ -17,8 +16,8 @@ describe("Png comparing", ({test, _}) => {
   });
 
   test("uses provided threshold", ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/orange.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/orange_changed.png");
+    let img1 = Png.IO.loadImage("test/test-images/orange.png");
+    let img2 = Png.IO.loadImage("test/test-images/orange_changed.png");
 
     let (_, diffPixels, diffPercentage) =
       Diff.compare(img1, img2, ~threshold=0.5, ());
@@ -27,8 +26,8 @@ describe("Png comparing", ({test, _}) => {
   });
 
   test("uses provided irgnore regions", ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/orange.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/orange_changed.png");
+    let img1 = Png.IO.loadImage("test/test-images/orange.png");
+    let img2 = Png.IO.loadImage("test/test-images/orange_changed.png");
 
     let (_diffOutput, diffPixels, diffPercentage) =
       Diff.compare(
@@ -46,22 +45,18 @@ describe("Png comparing", ({test, _}) => {
   });
 
   test("creates the right diff output image", ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/orange.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/orange_changed.png");
+    let img1 = Png.IO.loadImage("test/test-images/orange.png");
+    let img2 = Png.IO.loadImage("test/test-images/orange_changed.png");
 
     let (diffOutput, _, _) = Diff.compare(img1, img2, ());
 
-    let originalDiff =
-      PureC_IO.IO.loadImage("test/test-images/orange_diff.png");
+    let originalDiff = Png.IO.loadImage("test/test-images/orange_diff.png");
     let (diffMaskOfDiff, diffOfDiffPixels, diffOfDiffPercentage) =
       Diff.compare(originalDiff, diffOutput, ());
 
     if (diffOfDiffPixels > 0) {
-      PureC_IO.IO.saveImage(diffOutput, "test/test-images/diff-output.png");
-      PureC_IO.IO.saveImage(
-        diffMaskOfDiff,
-        "test/test-images/diff-of-diff.png",
-      );
+      Png.IO.saveImage(diffOutput, "test/test-images/diff-output.png");
+      Png.IO.saveImage(diffMaskOfDiff, "test/test-images/diff-of-diff.png");
     };
 
     expect.int(diffOfDiffPixels).toBe(0);
@@ -71,23 +66,20 @@ describe("Png comparing", ({test, _}) => {
   test(
     "creates the right diff output image with custom diff color",
     ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/orange.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/orange_changed.png");
+    let img1 = Png.IO.loadImage("test/test-images/orange.png");
+    let img2 = Png.IO.loadImage("test/test-images/orange_changed.png");
 
     let (diffOutput, _, _) =
       Diff.compare(img1, img2, ~diffPixel=(0, 255, 0), ());
 
     let originalDiff =
-      PureC_IO.IO.loadImage("test/test-images/orange_diff_green.png");
+      Png.IO.loadImage("test/test-images/orange_diff_green.png");
     let (diffMaskOfDiff, diffOfDiffPixels, diffOfDiffPercentage) =
       Diff.compare(originalDiff, diffOutput, ());
 
     if (diffOfDiffPixels > 0) {
-      PureC_IO.IO.saveImage(
-        diffOutput,
-        "test/test-images/diff-output-green.png",
-      );
-      PureC_IO.IO.saveImage(
+      Png.IO.saveImage(diffOutput, "test/test-images/diff-output-green.png");
+      Png.IO.saveImage(
         diffMaskOfDiff,
         "test/test-images/diff-of-diff-green.png",
       );
@@ -99,9 +91,9 @@ describe("Png comparing", ({test, _}) => {
 
   test("does not count anti-aliased pixels as different", ({expect, _}) => {
     let img1 =
-      PureC_IO_Bigarray.IO.loadImage("test/test-images/antialiasing-on.png");
+      Png.BigarrayIO.loadImage("test/test-images/antialiasing-on.png");
     let img2 =
-      PureC_IO_Bigarray.IO.loadImage("test/test-images/antialiasing-off.png");
+      Png.BigarrayIO.loadImage("test/test-images/antialiasing-off.png");
 
     let (_, diffPixels, diffPercentage) =
       AADiff.compare(
@@ -117,14 +109,14 @@ describe("Png comparing", ({test, _}) => {
   });
 
   test("Diff of mask and no mask are equal", ({expect, _}) => {
-    let img1 = PureC_IO.IO.loadImage("test/test-images/antialiasing-on.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/antialiasing-off.png");
+    let img1 = Png.IO.loadImage("test/test-images/antialiasing-on.png");
+    let img2 = Png.IO.loadImage("test/test-images/antialiasing-off.png");
 
     let (_, diffPixels, diffPercentage) =
       Diff.compare(img1, img2, ~outputDiffMask=false, ());
 
-    let img1 = PureC_IO.IO.loadImage("test/test-images/antialiasing-on.png");
-    let img2 = PureC_IO.IO.loadImage("test/test-images/antialiasing-off.png");
+    let img1 = Png.IO.loadImage("test/test-images/antialiasing-on.png");
+    let img2 = Png.IO.loadImage("test/test-images/antialiasing-off.png");
 
     let (_, diffPixelsMask, diffPercentageMask) =
       Diff.compare(img1, img2, ~outputDiffMask=true, ());
@@ -135,11 +127,9 @@ describe("Png comparing", ({test, _}) => {
 
   test("tests diffrent sized AA images", ({expect, _}) => {
     let img1 =
-      PureC_IO_Bigarray.IO.loadImage("test/test-images/antialiasing-on.png");
+      Png.BigarrayIO.loadImage("test/test-images/antialiasing-on.png");
     let img2 =
-      PureC_IO_Bigarray.IO.loadImage(
-        "test/test-images/antialiasing-off-small.png",
-      );
+      Png.BigarrayIO.loadImage("test/test-images/antialiasing-off-small.png");
 
     let (_, diffPixels, diffPercentage) =
       AADiff.compare(
